push on iAPC enabled

+++ Jenkinsfile
@@ -20,8 +20,10 @@ pipeline {
                     sh 'cf api https://api.scapp-console.swisscom.com'
                     sh 'cf auth iAPC-ATS $CF_PASSWORD'
                     sh 'cf target -o INI-DOS-FDN-ENB_BDD_Showcase -s Dev'
-                    sh 'rm package-lock.json'
-                    sh 'rm -r node_modules'
+                    if ( fileExists 'package-lock.json' ) {
+                       sh 'rm package-lock.json'
+                       sh 'rm -r node_modules'
+                    }
                     sh 'npm install'
                     sh 'cp serenity-cli-2.1.9-all.jar node_modules/@serenity-js/cache'
                     sh 'npm run build'
